Apply the backtrace filter on the result in the new api
Move the backtrace filter out of the LegacyApi module and apply it on
the result arguments of the new formatter api methods. Expect that the
backtrace line of the step already has been appended to the backtrace
of the result objects.
