+++ lib/cucumber/formatter/backtrace_filter.rb
@@ -1,0 +1,40 @@
+require 'cucumber/platform'
+
+
+module Cucumber
+  module Formatter
+
+    class BacktraceFilter
+      BACKTRACE_FILTER_PATTERNS = \
+      [/vendor\/rails|lib\/cucumber|bin\/cucumber:|lib\/rspec|gems\/|minitest|test\/unit|.gem\/ruby|lib\/ruby/]
+      if(::Cucumber::JRUBY)
+        BACKTRACE_FILTER_PATTERNS << /org\/jruby/
+      end
+      PWD_PATTERN = /#{::Regexp.escape(::Dir.pwd)}\//m
+
+      def initialize(exception)
+        @exception = exception
+      end
+
+      def exception
+        return @exception if ::Cucumber.use_full_backtrace
+        @exception.backtrace.each{|line| line.gsub!(PWD_PATTERN, "./")}
+
+        filtered = (@exception.backtrace || []).reject do |line|
+          BACKTRACE_FILTER_PATTERNS.detect { |p| line =~ p }
+        end
+
+        if ::ENV['CUCUMBER_TRUNCATE_OUTPUT']
+          # Strip off file locations
+          filtered = filtered.map do |line|
+            line =~ /(.*):in `/ ? $1 : line
+          end
+        end
+
+        @exception.set_backtrace(filtered)
+        @exception
+      end
+    end
+
+  end
+end
